chore: moving ReplicatorStateMapInstaller behind a trait
