+++ packages/cohort/src/replicator/utils.rs
@@ -1,4 +1,12 @@
+use std::{io::Error, sync::Arc};
+
+use log::info;
 use talos_suffix::SuffixItem;
+
+use crate::{
+    state::postgres::{data_access::PostgresApi, database::Database},
+    tx_batch_executor::BatchExecutor,
+};
 
 use super::{core::StatemapItem, suffix::ReplicatorSuffixItemTrait};
 
@@ -27,4 +35,17 @@ pub fn get_statemap_from_suffix_items<'a, T: ReplicatorSuffixItemTrait + 'a>(mes
         }
         None => acc,
     })
+}
+
+pub async fn statemap_install_handler(sm: Vec<StatemapItem>, db: Arc<Database>, version: Option<u64>) -> Result<bool, Error> {
+    info!("Last version ... {:#?} ", version);
+    info!("Original statemaps received ... {:#?} ", sm);
+
+    let mut manual_tx_api = PostgresApi { client: db.get().await };
+
+    let result = BatchExecutor::execute(&mut manual_tx_api, sm, version).await;
+
+    info!("Result on executing the statmaps is ... {result:?}");
+
+    Ok(result.is_ok())
 }
